Resolve many size_t to int compile warnings. Some formatting changes. Implementation of ds::math::distance is now the glm implementation

+++ src/ds/app/engine/engine.cpp
@@ -480,7 +480,7 @@ void Engine::loadNinePatchCfg(const std::string& filename) {
 	mData.mEngineCfg.loadNinePatchCfg(filename);
 }
 
-int Engine::getRootCount() const {
+size_t Engine::getRootCount() const {
 	return mRoots.size();
 }
